Add NFC TagIntentAppPreference System API
Add System APIs for setting a preference list by filtering package names
before sending NFC TAG intents.

ACTION_NDEF_DISCOVERED, ACTION_TECH_DISCOVERED and ACTION_TAG_DISCOVERED
will not be dispatched to an Activity if its package is disallowed.

An app is added to the preference list with the allowed flag set to true
when a Tag intent is dispatched to the package for the first time.
The System APIs are called by settings to note that the user wants to
change this default preference.

Bug: 244272155
Test: atest NfcFeatureFlagTest
Test: atest android.permission.cts.NfcPermissionTest